added logic to add start and end offset to all refactorings to implement the highlighting where needed

+++ src/main/java/models/refactorings/ChangeAttributeType.java
@@ -5,8 +5,8 @@ public class ChangeAttributeType extends RefactoringData {
 
     public ChangeAttributeType(String... attributes) {
         this.oldType = attributes[0];
-        this.startOffset = Integer.parseInt(attributes[1]);
-        this.endOffset = Integer.parseInt(attributes[2]);
+        this.startOffset = Long.parseLong(attributes[1]);
+        this.endOffset = Long.parseLong(attributes[2]);
     }
 
     @Override
